Create separate folder for calculator component
