+++ src/components/App.js
@@ -1,5 +1,5 @@
 import { Route, Switch } from 'react-router-dom';
-import Calculator from './Calculator';
+import Calculator from './calculator/Calculator';
 import '../style/App.css';
 
 const App = () => (
