+++ src/components/calculator/Button.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import '../style/App.css';
+import '../../style/App.css';
 
 const Button = ({ name, onClick, className }) => (
   <button type="button" onClick={() => onClick(name)} className={className}>
