+++ src/components/calculator/Calculator.js
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import Display from './Display';
 import ButtonPanel from './ButtonPanel';
-import calculate from '../logic/calculate';
-import '../style/App.css';
-import Navbar from './shared/Navbar';
+import calculate from '../../logic/calculate';
+import '../../style/App.css';
+import Navbar from '../shared/Navbar';
 
 const Calculator = () => {
   const [state, setState] = useState({ total: '', next: '', operation: '' });
