+++ src/components/shared/Navbar.js
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import '../../style/App.css';
 
 const Navbar = () => (
   <div>
