+++ src/style/App.css
@@ -51,10 +51,3 @@
   border-radius: 15px;
   padding: 0 10px;
 }
-
-#root {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
